+++ src/main/java/com/groom/orbit/member/auth/app/oauth/kakao/KakaoLoginParams.java
@@ -1,20 +1,23 @@
-package com.groom.orbit.member.auth.app.kakao;
+package com.groom.orbit.member.auth.app.oauth.kakao;
 
 import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.util.MultiValueMap;
+
+import com.groom.orbit.member.auth.app.oauth.OAuthLoginParams;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor
-public class KakaoReissueParams {
+public class KakaoLoginParams implements OAuthLoginParams {
 
-  private String refreshToken;
+  private String authorizationCode;
 
+  @Override
   public MultiValueMap<String, String> makeBody() {
     MultiValueMap<String, String> body = new LinkedMultiValueMap<>();
-    body.add("refresh_token", refreshToken);
+    body.add("code", authorizationCode);
     return body;
   }
 }
