+++ src/main/java/com/groom/orbit/member/auth/app/oauth/kakao/KakaoReissueParams.java
@@ -1,23 +1,20 @@
-package com.groom.orbit.member.auth.app.kakao;
+package com.groom.orbit.member.auth.app.oauth.kakao;
 
 import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.util.MultiValueMap;
-
-import com.groom.orbit.member.auth.app.oauth.OAuthLoginParams;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor
-public class KakaoLoginParams implements OAuthLoginParams {
+public class KakaoReissueParams {
 
-  private String authorizationCode;
+  private String refreshToken;
 
-  @Override
   public MultiValueMap<String, String> makeBody() {
     MultiValueMap<String, String> body = new LinkedMultiValueMap<>();
-    body.add("code", authorizationCode);
+    body.add("refresh_token", refreshToken);
     return body;
   }
 }
