+++ src/main/java/com/groom/orbit/member/auth/controller/OAuthController.java
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.RestController;
 import com.groom.orbit.common.exception.BaseResponse;
 import com.groom.orbit.member.auth.app.AuthService;
 import com.groom.orbit.member.auth.app.dto.LoginResponseDto;
-import com.groom.orbit.member.auth.app.kakao.KakaoLoginParams;
+import com.groom.orbit.member.auth.app.oauth.kakao.KakaoLoginParams;
 
 import lombok.RequiredArgsConstructor;
